marketplace/mdc/requirements.txt
Django==1.8.14
djangorestframework==3.1.0
requests==2.5.3
MySQL-python==1.2.5
python-dateutil
django-extensions
django-rest-swagger
django-cors-headers
